backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlmodel==0.0.16
pydantic==2.5.0
pydantic-settings==2.1.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
better-exceptions==0.3.2
python-dotenv==1.0.0
aiosqlite==0.19.0
alembic==1.13.1